redux reconfigured
