+++ package.json
@@ -15,6 +15,7 @@
     "deep-equal": "^2.0.5",
     "next": "^10.0.8",
     "next-images": "^1.7.0",
+    "next-redux-wrapper": "^6.0.2",
     "public-ip": "^4.0.3",
     "pure-react-carousel": "^1.27.5",
     "query-string": "^6.13.7",
@@ -26,6 +27,7 @@
     "react-portal": "^4.2.1",
     "react-redux": "^7.2.1",
     "react-visibility-sensor": "^5.1.1",
+    "redux-devtools-extension": "^2.13.9",
     "uuid": "^8.3.1"
   },
   "scripts": {
